+++ app/src/main/java/com/example/coffetech/viewmodel/Transaction/EditTransactionViewModel.kt
@@ -96,7 +96,7 @@ class EditTransactionViewModel : ViewModel() {
         transactionTypeName: String,
         transactionCategoryName: String,
         description: String?,
-        value: Int,
+        value: Long,
         transactionDate: String
     ) {
         // Establecer los valores iniciales
@@ -277,7 +277,7 @@ class EditTransactionViewModel : ViewModel() {
                         transaction_type_name = _selectedTransactionType.value,
                         transaction_category_name = _selectedTransactionCategory.value,
                         description = _descripcion.value,
-                        value = _valor.value.toInt(),
+                        value = _valor.value.toLong(),
                         transaction_date = _fecha.value
                     )
